move: 회고 생성 이벤트 클래스 이벤트 모듈로 이동
