+++ layer-api/src/main/java/org/layer/domain/retrospect/service/RetrospectService.java
@@ -28,7 +28,7 @@ import org.layer.domain.space.entity.Space;
 import org.layer.domain.space.entity.Team;
 import org.layer.domain.space.repository.MemberSpaceRelationRepository;
 import org.layer.domain.space.repository.SpaceRepository;
-import org.layer.discord.event.CreateRetrospectEvent;
+import org.layer.event.retrospect.CreateRetrospectEvent;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
