+++ layer-event/src/main/java/org/layer/event/retrospect/CreateRetrospectEvent.java
@@ -1,4 +1,4 @@
-package org.layer.discord.event;
+package org.layer.event.retrospect;
 
 import java.time.LocalDateTime;
